pyproject.toml
[project]
name = "validate-actions"
version = "1.0.3"
description = "CLI for github actions workflow validation"
authors = [
    {name = "konradhorber",email = "konrad.horber@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer (>=0.15.2,<0.16.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "rich (>=13.0.0,<14.0.0)",
]

[tool.poetry]
packages = [
    { include="validate_actions" }
]

[tool.poetry.scripts]
validate-actions = "validate_actions.main:app"

[tool.pytest.ini_options]
pythonpath = [
  "validate_actions"
]

[tool.coverage.run]
omit = [
    "scripts/*",
    "tests/*",
    "*/__init__.py",
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
types-pyyaml = "^6.0.12.20250402"
coverage = "^7.8.0"
black = "^24.0.0"
flake8 = "^7.0.0"
isort = "^5.13.0"
mypy = "^1.16.1"
types-requests = "^2.32.4.20250611"
sphinx = "^8.2.3"

[tool.black]
line-length = 99
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 99

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
